re-organize testutil code in different packages
