+++ dynamodb/api.go
@@ -2,13 +2,9 @@ package dynamodb
 
 import (
 	"context"
-	"errors"
-	"strings"
 
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb"
-	"github.com/aws/aws-sdk-go-v2/service/dynamodb/types"
-	"github.com/aws/smithy-go"
-	"github.com/aws/smithy-go/transport/http"
+	"github.com/ln80/pii/dynamodb/testutil"
 )
 
 // ClientAPI presents an interface for a sub-part of the AWS Dynamodb client service:
@@ -22,26 +18,6 @@ type ClientAPI interface {
 	DeleteItem(ctx context.Context, params *dynamodb.DeleteItemInput, optFns ...func(*dynamodb.Options)) (*dynamodb.DeleteItemOutput, error)
 }
 
-// isConditionCheckFailure checks if the given error is an aws error that expresses a conditional failure exception.
-// It works seamlessly in both single write and within a transaction operation.
-func isConditionCheckFailure(err error) bool {
-	if strings.Contains(err.Error(), "ConditionalCheckFailedException") {
-		return true
-	}
-	var oe *smithy.OperationError
-	if errors.As(err, &oe) {
-		var re *http.ResponseError
-		if errors.As(err, &re) {
-			var tce *types.TransactionCanceledException
-			if errors.As(err, &tce) {
-				for _, reason := range tce.CancellationReasons {
-					if *reason.Code == "ConditionalCheckFailed" {
-						return true
-					}
-				}
-			}
-		}
-	}
-
-	return false
-}
+// CreateTable is an alias of CreateTable func defined in testutil package.
+// Mainly used for test and local dev purposes.
+var CreateTable = testutil.CreateTable
