+++ kms/key_engine_test.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"testing"
 
+	kms_testutil "github.com/ln80/pii/kms/testutil"
 	"github.com/ln80/pii/memory"
 	"github.com/ln80/pii/testutil"
 )
@@ -13,7 +14,7 @@ func TestKeyEngine(t *testing.T) {
 
 	originEng := memory.NewKeyEngine()
 
-	testutil.WithKMSKey(t, func(kmsvc interface{}, key string) {
+	kms_testutil.WithKMSKey(t, func(kmsvc interface{}, key string) {
 		resolver := NewStaticKMSKeyResolver(key)
 
 		eng := NewKMSWrapper(kmsvc.(ClientAPI), resolver, originEng)
